Move `src/stories` dir to `src/dev/stories`
- `src/dev` should contain all dev related stuff
